Remove tabs, general cleanup
