+++ components/index.js
@@ -1,7 +1,6 @@
 /**
  *  Point of contact for component modules
  */
-export EditorTranspiler from './EditorTranspiler';
 export Phone from './Phone';
 export Player from './Player';
 export EditorPlayer from './EditorPlayer';
